6465302 integrate Java DTrace API tests with DTrace test suite 6467526 test suite: sched/tst.stackdepth.d fails in empty stack case

+++ usr/src/cmd/dtrace/test/tst/common/ustack/tst.depth.ksh
@@ -25,8 +25,13 @@
 #
 #ident	"%Z%%M%	%I%	%E% SMI"
 
+if [ $# != 1 ]; then
+	echo expected one argument: '<'dtrace-path'>'
+	exit 2
+fi
+
 file=out.$$
-dtrace=/usr/sbin/dtrace
+dtrace=$1
 
 rm -f $file
